refactor: Move Units to org.opentripplanner.framework.model
