+++ src/main/java/org/opentripplanner/routing/api/request/preference/ElevatorPreferences.java
@@ -4,8 +4,8 @@ import java.io.Serializable;
 import java.util.Objects;
 import java.util.function.Consumer;
 import org.opentripplanner.framework.model.Cost;
+import org.opentripplanner.framework.model.Units;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
-import org.opentripplanner.routing.api.request.framework.Units;
 
 /**
  *  TODO: how long does it /really/ take to  an elevator?
