+++ src/main/java/org/opentripplanner/routing/api/request/preference/ItineraryFilterPreferences.java
@@ -2,11 +2,11 @@ package org.opentripplanner.routing.api.request.preference;
 
 import java.util.Objects;
 import java.util.function.Consumer;
+import org.opentripplanner.framework.model.Units;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
 import org.opentripplanner.routing.algorithm.filterchain.api.TransitGeneralizedCostFilterParams;
 import org.opentripplanner.routing.api.request.framework.DoubleAlgorithmFunction;
 import org.opentripplanner.routing.api.request.framework.RequestFunctions;
-import org.opentripplanner.routing.api.request.framework.Units;
 
 /**
  * Group by Similarity filter parameters. See the configuration for documentation of each field.
