+++ src/main/java/org/opentripplanner/routing/api/request/preference/RaptorPreferences.java
@@ -10,11 +10,11 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.function.Consumer;
 import javax.annotation.Nullable;
+import org.opentripplanner.framework.model.Units;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
 import org.opentripplanner.raptor.api.model.SearchDirection;
 import org.opentripplanner.raptor.api.request.Optimization;
 import org.opentripplanner.raptor.api.request.RaptorProfile;
-import org.opentripplanner.routing.api.request.framework.Units;
 
 /**
  * Set of optimizations to use with Raptor. These are available here for testing purposes.
