+++ src/main/java/org/opentripplanner/routing/api/request/preference/TransferOptimizationPreferences.java
@@ -5,9 +5,9 @@ import static org.opentripplanner.framework.lang.DoubleUtils.doubleEquals;
 import java.io.Serializable;
 import java.util.Objects;
 import org.opentripplanner.framework.application.OTPFeature;
+import org.opentripplanner.framework.model.Units;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
 import org.opentripplanner.routing.algorithm.transferoptimization.api.TransferOptimizationParameters;
-import org.opentripplanner.routing.api.request.framework.Units;
 
 /**
  * See {@link TransferOptimizationParameters} for documentation on this class.
