+++ src/main/java/org/opentripplanner/routing/api/request/preference/TransferPreferences.java
@@ -7,9 +7,9 @@ import java.io.Serializable;
 import java.util.Objects;
 import java.util.function.Consumer;
 import org.opentripplanner.framework.model.Cost;
+import org.opentripplanner.framework.model.Units;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
 import org.opentripplanner.routing.algorithm.transferoptimization.api.TransferOptimizationParameters;
-import org.opentripplanner.routing.api.request.framework.Units;
 
 /**
  * Parameters for doing transfers between transit legs.
