+++ src/main/java/org/opentripplanner/routing/api/request/preference/WalkPreferences.java
@@ -6,8 +6,8 @@ import java.io.Serializable;
 import java.util.Objects;
 import java.util.function.Consumer;
 import org.opentripplanner.framework.model.Cost;
+import org.opentripplanner.framework.model.Units;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
-import org.opentripplanner.routing.api.request.framework.Units;
 
 /**
  * The walk preferences contain all speed, reluctance, cost and factor preferences for walking
