+++ src/main/java/org/opentripplanner/routing/api/request/preference/WheelchairPreferences.java
@@ -5,8 +5,8 @@ import static org.opentripplanner.routing.api.request.preference.AccessibilityPr
 import java.io.Serializable;
 import java.util.Objects;
 import java.util.function.Consumer;
+import org.opentripplanner.framework.model.Units;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
-import org.opentripplanner.routing.api.request.framework.Units;
 
 /**
  * See the configuration for documentation of each field.
